Split modifyInventory component
